PCI scan failing on the existence of CFID cookie when using jsessionid
A PCI scan of our website revealed that CFID and CFTOKEN cookies were being set even though we have our ColdFusion server set to use J2EE session variables (jsessionid).  It is my understanding that when you have the server configured to use jsessionid the CFID and CFTOKEN values are not used and are no longer necessary.  They are known to be predictable.

NOTE: I'm not positive if the same applies to Lucee/Railo although I would presume that to be the case. I will leave that up to you guys to confirm.  We are using Adobe CF.

NOTE: This code only runs if you have set the secureCookies value to "true" in the settings.ini.cfm file.  With newer versions of ColdFusion (Adobe at least, version 10+) you can set the session cookies to be secure using the CF administrator.